feat: complete all the elements in desktop 71, refactor breadcrumb component, adds open sans font
